% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AucBinormal.R
\name{auc.binormal}
\alias{auc.binormal}
\title{Area Under the Binormal ROC Curve}
\usage{
\method{auc}{binormal}(fit.obj, a, b, vcov, level = 0.95, ...)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{f.1}{Regression formula for parameters. For example, \code{a~x+z} and \code{b~1}. Order does not matter.}

\item{f.2}{Same as \code{f.1}.}

\item{f.3}{Same as \code{f.1}.}

\item{f.4}{Same as \code{f.1}.}

\item{model}{Model specification. Options are \code{"binormal"}, \code{"Bilogistic"}, \code{"as.defined"}. The last option is defined by the next 4 parameters.}

\item{dist.0}{Distribution F (e.g. negatives or group 0). Options are \code{"normal"}, \code{"logistic"}, \code{"gamma"}.}

\item{dpar.0}{Parameter list for distribution F}

\item{dist.1}{Distribution G (e.g. positives or group 1).}

\item{dpar.1}{Parameter list for distribution G}

\item{parms}{values for parameters \code{c(a, b, theta.0, theta.1)}}
}
\value{
A list with the following elements:

\describe{
\item{\code{coef}}{Maximum likelihood estimates of the parameter vector.}
\item{\code{fit.var}}{Variance-covariance matrix for coefficients}
}
}
\description{
Comptues the analytical estimate of the AUC for a binormal model, along with a delta-method standard error.
}
\details{
Performs a maxmum likelihood fit of the ROC model.
}
\examples{

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{set.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
