% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AucGeneric.R
\name{auc}
\alias{auc}
\title{Generic function for AUC computations}
\usage{
auc(x, ...)
}
\arguments{
\item{x}{An ROC or Curve object.}
}
\value{

}
\description{

}
\details{

}
\examples{

}
\references{
Add references here
}
\seealso{
\url{auc.curve} \code{\link{auc.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
