% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AucRoc.R
\name{auc.roc}
\alias{auc.roc}
\title{Empirical area under the ROC curve (AUC)}
\usage{
\method{auc}{roc}(roc.obj, level = 0.95, digits = 4, ...)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{level}{Confidence interval coverage level. Default is \code{95\%}.}

\item{digits}{Rounding accuracy for display. Default is \code{4}.}
}
\value{
Returns the empirical estimate of the area under the ROC curve, along with an approximate and robust confidence intervals:

\describe{
\item{\code{auc}}{Area under the ROC curve (AUC).}
\item{\code{ci.mwu}}{Confidence interval based on Mann-Whitney U-statistic.}
\item{\code{ci.rob}}{Confidence interval based on distribution robust SEs.}
}
}
\description{
This function computes the area under the empirical ROC curve.
}
\details{
Computes the empirical area under the ROC curve. This is also known as the Mann-Whitney-U-statistic and c-statistic. The reported AUC is actually \code{max(AUC, 1-AUC)}. This ensures the estimate is always greater than 0.5. Approximate and robust standard errors are used to obtain CIs.
}
\examples{

## Fake ROC data for examples 
y.0 <- round(rnorm(100, mean=0, sd=1)   , 1) 
y.1 <- round(rnorm(100, mean=1, sd=1.1) , 1)
myroc <- set.roc(score=y.0, grp=y.1, as.groups=TRUE)

## Area under empirical ROC curve
auc(myroc)

## Empirical AUC from smoothing function
auc(smooth(myroc, adj=0.01))

}
\references{
Blume 2009 JSPI \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2631183/} and \url{https://rdrr.io/cran/asht/man/wmwTest.html}
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{auc.curve}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
