% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{cost}
\alias{cost}
\title{Calculate a cost function.}
\usage{
cost(fpr, tpr, fp.weight = 1)
}
\arguments{
\item{fpr}{A numeric vector, representing the false positive rate (FPR)
of points on a ROC curve.}

\item{tpr}{A numeric vector, representing the true positive rate (TPR) of
points on a ROC curve, assumed to be corresponding to those represented
in \code{fpr}.}

\item{fp.weight}{A weight (numeric value) representing the relative
importance (cost) of a false positive, relative to a false negative.
So a value of 2 would indicate a false positive is twice as `bad'
as a false negative, a value of 0.5 would indicate it is half as bad.}
}
\value{
A vector of costs.
}
\description{
Given a receiver operating characteristic (ROC) curve,
calculates the cost function with given weights on false positives
and false negatives respectively.
}
\details{
Arguments \code{fpr} and \code{tpr} must contain only values between
0 and 1, without any NA or NaN values, and must be the same length.
Otherwise, \code{cost} will throw an appropriate error.
}
\examples{
roc.df = roc(c(1, 2, 3, 4, 4, 5), c(FALSE, TRUE, FALSE, FALSE, TRUE, TRUE))
cost(roc.df$FPR, roc.df$TPR)
cost(roc.df$FPR, roc.df$TPR, 2)
}

