% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveGeneric.R
\name{curve}
\alias{curve}
\title{Generic function for obtaining ROC curves from fitted ROC objects}
\usage{
curve(fit.obj, ...)
}
\arguments{
\item{x}{An ROC fitted object.}
}
\value{

}
\description{

}
\details{

}
\examples{

}
\references{
Add references here
}
\seealso{
\url{smooth.default} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{ROC}
\keyword{curve,}
\keyword{density}
\keyword{smooth,}
