% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutPlot.R
\name{cutplot}
\alias{cutplot}
\title{Plot histograms of ROC data and estimated densities}
\usage{
cutplot(
  roc.obj,
  fit.obj = NULL,
  at,
  adj = 0.5,
  ref = TRUE,
  smooth = FALSE,
  cols = c("dodgerblue3", "chocolate3"),
  lwds = c(2.5, 2.5),
  show.legend = TRUE,
  show.auc = TRUE,
  xlab = "Score",
  plot = TRUE,
  ...
)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{ref}{Flag to replace smoothed sens/spec reference with empirical estimates}

\item{lwds}{Density line widths for group 0 and group 1. Default is \code{c(3,3)}.}

\item{show.legend}{Flag to display default legend. Default is \code{TRUE}.}

\item{xlab}{Lable for x-axis.}

\item{plot}{Flag to turn plot off}

\item{curve.obj}{Curve object. For example, from \code{smooth} function.}

\item{breaks}{Histogram breaks. Default is \code{"Scott"} but \code{"FD"} is also a good option.}

\item{hcols}{Distogram colors for group 0 and group 1. Default is \code{c("steelblue1", "darkorange")}.}

\item{dcols}{Density line colors for group 0 and group 1. Default is \code{c("dodgerblue3", "chocolate3")}.}

\item{title}{Title}

\item{blty}{Line type for smoothed densities when plotted with fitted densities (solid line).}

\item{blwd}{Line width for smoothed densities when plotted with fitted densities (solid line).}
}
\value{
A list of sen/spec tradeoff that is plotted
}
\description{
This function plots histograms of the ROC and associated smooth or fitted densities.
}
\details{
Details?
}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{plot.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{Histogram}
\keyword{ROC}
\keyword{curve,}
