% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGeneric.R
\name{fit}
\alias{fit}
\title{Generic function for ROC model fitting}
\usage{
fit(roc.obj, ...)
}
\arguments{
\item{x}{An ROC object.}
}
\description{
Generic function for ROC model fitting
}
\references{
Add references here
}
\seealso{
\url{fit.roc}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
