% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitRoc.R
\name{fit.roc}
\alias{fit.roc}
\title{Maximum Likelihood Fit for ROC model}
\usage{
\method{fit}{roc}(
  roc.obj,
  f.1 = NULL,
  f.2 = NULL,
  f.3 = NULL,
  f.4 = NULL,
  model = "binormal",
  dist.0 = NA,
  dist.1 = NA,
  dpar.0 = NA,
  dpar.1 = NA,
  start.parms = NA,
  method = "BFGS",
  ...
)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{f.1}{Regression formula for parameters. For example, \code{a~x+z} and \code{b~1}. Order does not matter.}

\item{f.2}{Same as \code{f.1}.}

\item{f.3}{Same as \code{f.1}.}

\item{f.4}{Same as \code{f.1}.}

\item{model}{Model specification. Options are \code{"binormal"}, \code{"bilogistic"}, \code{"other"}. The last option is defined by the next 4 parameters.}

\item{dist.0}{Distribution F (e.g. negatives or group 0). Options are \code{"normal"}, \code{"logistic"}, \code{"gamma"}.}

\item{dist.1}{Distribution G (e.g. positives or group 1).}

\item{dpar.0}{Parameter list for distribution F}

\item{dpar.1}{Parameter list for distribution G}

\item{method}{Optimization algorithm passed to \code{optim}.}

\item{parms}{Values for parameters \code{c(a, b, theta.0, theta.1)}}
}
\value{
A list with the following elements:

\describe{
\item{\code{coef}}{Maximum likelihood estimates of the parameter vector.}
\item{\code{vcov}}{Variance-covariance matrix for coefficients}
\item{\code{more}}{more here..!}
\item{\code{adjr2}}{slightly modified adjusted R^2}
}
}
\description{
This function will maximize the likelihood function for ROC parameters. Used to estimate the ROC parameters.
}
\details{
Performs a maxmum likelihood fit of the ROC model.
}
\examples{

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{set.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
