% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedRoc.R
\name{fitted.roc}
\alias{fitted.roc}
\title{Compute fitted values for an ROC model}
\usage{
\method{fitted}{roc}(parms, roc.obj, mm.list, ...)
}
\arguments{
\item{parms}{values for parameters \code{c(a, b, theta.0, theta.1)}}

\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{mm.list}{List of Model Matrix and numer of parameters for each formula. Obtained from \code{get.modelmatrix}.}
}
\value{
list of fittedvalues (NEED MORE HERE)
}
\description{
This function will compute the fitted values for an ROC model.
}
\details{
Computes fitted values, fitted variance, and MSE for ROC model.
}
\examples{

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
