% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFormula.R
\name{get.formula}
\alias{get.formula}
\title{Parse and organize the 4 ROC parameter formulas}
\usage{
get.formula(f.1 = NULL, f.2 = NULL, f.3 = NULL, f.4 = NULL, ...)
}
\arguments{
\item{f.1}{Regression formula for parameters where the dependent varaiable is one of a, b, t0, or t1. For example, \code{a~x+z} and \code{b~1}.}

\item{f.2}{Same as \code{f.1}.}

\item{f.3}{Same as \code{f.1}.}

\item{f.4}{Same as \code{f.1}.}
}
\value{
ROC model formulas for ROC parameters a, b, theta.0, and theta.1.
}
\description{
This function takes the 4 formula inputs and organizes and formats them for the ROC analysis.
}
\details{
Formulas can be in any order. Dependent variable in each formula must be one of a, b, theta.0, or theta.1. Duplicate dependent variables not allowed.
}
\examples{
f.1=a~1 ; f.2=t1~size ; f.3=NULL ; f.4=b~age
get.formula(f.1, f.2, f.3, f.4)

f.1=a~1 ; f.2=t1~size ; f.3=NULL ; f.4=b~age
get.formula(f.1, f.2, f.3, f.4)

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
