% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetModelmatrix.R
\name{get.modelmatrix}
\alias{get.modelmatrix}
\title{Obtain 4 ROC model matricies and ranks}
\usage{
get.modelmatrix(roc.obj, f.a, f.b, f.t0, f.t1, ...)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{f.a}{Regression formula for a obtained from \code{get.formula} function.}

\item{f.b}{Regression formula for b obtained from \code{get.formula} function.}

\item{f.t0}{Regression formula for t0 obtained from \code{get.formula} function.}

\item{f.t1}{Regression formula for t1 obtained from \code{get.formula} function.}
}
\value{
Model matrix for the 4 ROC formulas for a, b, theta.0, theta.1 and the parameter count for each formula statment.
}
\description{
This function compute the model matrix for the 4 ROC formulas and computes rank etc.
}
\details{
Formulas are specific, can not be \code{NULL} nor \code{NA}.
}
\examples{

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
