% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistRoc.R
\name{hist.roc}
\alias{hist.roc}
\title{Plot histograms of ROC data and estimated densities}
\usage{
\method{hist}{roc}(
  roc.obj,
  curve.obj = NULL,
  fit.obj = NULL,
  breaks = "Scott",
  hcols = c("steelblue1", "darkorange"),
  dcols = c("dodgerblue3", "chocolate3"),
  lwds = c(3, 3),
  show.legend = TRUE,
  title = "Density Estimate",
  xlab = "Score",
  blty = 3,
  blwd = 2,
  ...
)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{curve.obj}{Curve object. For example, from \code{smooth} function.}

\item{breaks}{Histogram breaks. Default is \code{"Scott"} but \code{"FD"} is also a good option.}

\item{hcols}{Distogram colors for group 0 and group 1. Default is \code{c("steelblue1", "darkorange")}.}

\item{dcols}{Density line colors for group 0 and group 1. Default is \code{c("dodgerblue3", "chocolate3")}.}

\item{lwds}{Density line widths for group 0 and group 1. Default is \code{c(3,3)}.}

\item{show.legend}{Flag to display default legend. Default is \code{TRUE}.}

\item{title}{Title}

\item{xlab}{Lable for x-axis.}

\item{blty}{Line type for smoothed densities when plotted with fitted densities (solid line).}

\item{blwd}{Line width for smoothed densities when plotted with fitted densities (solid line).}
}
\value{

}
\description{
This function plots histograms of the ROC and associated smooth or fitted densities.
}
\details{
Details?
}
\examples{

## Fake ROC data for examples 
y.0 <- round(rnorm(100, mean=0, sd=1)   , 1) 
y.1 <- round(rnorm(100, mean=1.75, sd=1.1) , 1)
myroc <- set.roc(score=y.0, grp=y.1, as.groups=TRUE)

myroc.smooth <- smooth(myroc)

## Histogram of data and smoothed densities
hist(myroc)
hist(myroc, myroc.smooth)

## More smooth, alternate colors, forced breaks
hist(myroc, smooth(myroc,adj=1), hcols=c("dodgerblue3","firebrick"), dcols=c("darkblue","darkred"), seq(-4, 6, length.out = 21))


}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{plot.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{Histogram}
\keyword{ROC}
\keyword{curve,}
