% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikeRoc.R
\name{like.roc}
\alias{like.roc}
\title{Compute Likelihood for ROC model}
\usage{
like.roc(
  parms,
  roc.obj,
  mm.list,
  model = "binormal",
  dist.0 = NA,
  dist.1 = NA,
  dpar.0 = NA,
  dpar.1 = NA,
  ...
)
}
\arguments{
\item{parms}{values for parameters \code{c(a, b, theta.0, theta.1)}}

\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{mm.list}{List of Model Matrix and numer of parameters for each formula. Obtained from \code{get.modelmatrix}.}

\item{model}{Model type. Options are \code{"binormal"}, \code{"bilogistic"}, \code{"other"}. The last option is defined by the next 4 parameters.}

\item{dist.0}{Distribution F (e.g. negatives or group 0). Options are \code{"normal"}, \code{"logistic"}, \code{"gamma"}.}

\item{dist.1}{Distribution G (e.g. positives or group 1).}

\item{dpar.0}{Parameter list for distribution F}

\item{dpar.1}{Parameter list for distribution G}
}
\value{
negative log-likeliihood
}
\description{
This function will compute the likelihood for a given ROC model and data. Used for maximum likelihood computations.
}
\details{
Computes log likelihood for ROC model.
}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
