% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRoc.R
\name{plot.roc}
\alias{plot.roc}
\title{Plot Empirical ROC points and confidence region}
\usage{
\method{plot}{roc}(
  roc.obj,
  show.plot = TRUE,
  show.points = TRUE,
  roc.line = "none",
  ci.region = FALSE,
  flip.xaxis = FALSE,
  points.col = "black",
  points.pch = 20,
  line.col = "black",
  line.type = 1,
  line.lwd = 1.5,
  ci.border = NA,
  ci.color = "aliceblue",
  ...
)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{show.plot}{Flag for new plot. Default is \code{TRUE}.}

\item{show.points}{Flag to plot empirical ROC points. Default is \code{TRUE}.}

\item{roc.line}{Options for connecting empirical ROC points are \code{'step.fun'}, \code{'linear'}, \code{'none'}. . Default is \code{'none'}. See details.}

\item{ci.region}{flag for plotting pointwise CI region based on empirical pts. Default is \code{FALSE}.}

\item{flip.xaxis}{Flag to flip x-axis. Will show specificity instead of false positive rate (1-specificity). Default is \code{FALSE}.}

\item{points.col}{Color for points. Default is \code{'black'}.}

\item{points.pch}{Plotting symbol for points. Default is \code{pch=20}.}

\item{line.col}{color for ROC line. Default is \code{'black'}.}

\item{line.type}{line style for line for ROC line. Default is \code{lty=1}.}

\item{line.lwd}{line width for line for ROC line. Default is \code{lwd=2}.}

\item{ci.border}{flag for drawing border lines of CI region. Default is \code{NA}.}

\item{ci.color}{set color of CI region. Default is \code{'aliceblue'}.}
}
\value{

}
\description{
This function plots empirical ROC points and their pointwise confidence region.
}
\details{
The ROC line can be set to either a step function, linear interpolation between the points (connect the dots), or none. The linear interpolation corresponds with empirical estimates of the area under the ROC curve, but can be misleading when there are few categories or many ties in the data.

The confidence interval region is essentially the union of the confidence regions over all empirical points. Recall that each point has uncertainty along both the y- and x-axes. The CI limits are obtained when the ROC object is created with the \code{set.roc} function. If these were not computed, the function will not accept the \code{ci.region=TRUE} setting.
}
\examples{

## Fake ROC data for examples (with ties)
y.0 <- round( rnorm(100, mean=0, sd=1)   , 1)
y.1 <- round( rnorm(100, mean=1, sd=1.1) , 1)
myroc <- set.roc(score=y.0, grp=y.1, as.groups=TRUE)

## Various plots
plot(myroc)
plot(myroc, flip.xaxis=TRUE, ci.region=TRUE)
plot(myroc, roc.line="linear")
plot(myroc, roc.line="step.fun", line.col="green", show.plot=FALSE)

## CI region shows the span of the CIs around the empirical ROC points
plot(myroc, roc.line="linear", line.col="blue", points.col="blue", show.points=TRUE, ci.region=TRUE, ci.border="black")
points(myroc$fp.ci[,1], myroc$tp, col='hotpink',pch=20) 		## lower CI limit on FPR
points(myroc$fp, myroc$tp.ci[,2], col='purple',pch=20) 		## Upper CI limit on TPR
points(myroc$fp.ci[,2], myroc$tp, col='hotpink',pch=20)		## Upper CI limit on FPR
points(myroc$fp, myroc$tp.ci[,1], col='purple',pch=20)		## lower CI limit on TPR
legend('bottomright', c("Empirical ROC", "Edge of CI region", "FP CI limit", "TP CI limit"), col=c("blue", "black", "hotpink", "purple"), pch=c(20,NA,20,20), lty=c(1,1,NA,NA), bty='n')

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
