% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictFit.R
\name{predict.fitobj}
\alias{predict.fitobj}
\title{Predict ROC intercept (a) and slope (b) for covariate pattern}
\usage{
\method{predict}{fitobj}(fit.obj, at, level = 0.95, ...)
}
\arguments{
\item{fit.obj}{FITOBJ object from \code{fit.roc} function.}

\item{at}{dataframe of covariate profiles}

\item{level}{Confidence Interval level}
}
\value{
A list with the following elements:

\describe{
\item{\code{a}}{Vector of a estimates}
\item{\code{b}}{Vector of b estimates}
\item{\code{vcov}}{Variance-covariance matrix for ROC coefficients}
\item{\code{auc}}{Vector of auc estimates}
\item{\code{auc.lo}}{Vector of lower auc CI estimates}
\item{\code{auc.hi}}{Vector of upper auc CI estimates}
\item{\code{t0}}{Vector of t0 estimates}
\item{\code{t1}}{Vector of t1 estimates}
}
}
\description{
This function computes the ROC intercept (a) and slop (b) for the given covaraite patterns. Fitted values for outcomes are already computed in rocfit object.
}
\details{
Computes a contrast based on the a maxmum likelihood fit of the ROC model.
}
\examples{

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{set.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC,}
\keyword{ROC}
\keyword{curve,}
\keyword{predict}
