% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetRoc.R
\name{set.roc}
\alias{set.roc}
\title{Create an ROC object}
\usage{
set.roc(
  score,
  grp,
  x = NULL,
  as.groups = FALSE,
  level = 0.95,
  flip.grps = TRUE,
  get.ci = TRUE,
  center = TRUE,
  ...
)
}
\arguments{
\item{score}{Outcome vector.}

\item{grp}{Grouping vector (must be 2 exactly levels).}

\item{x}{Covariate matrix / dataframe (one column per covariate).}

\item{as.groups}{Indicator for alternate data format (\code{TRUE} or \code{FALSE}). See Details.}

\item{level}{Confidence interval level (default is \code{95\%}).}

\item{flip.grps}{Indicator if group with largest mean should be labeled as group 1. See details.}

\item{get.ci}{Flag to compute confidence intervals for the empirical ROC points.}

\item{center}{Flag to center all non-numeric and non-binary (0-1) variables. Default is \code{TRUE}.}
}
\value{
An ROC object (list) is returned with the following elements:

\describe{
\item{\code{fp}}{Vector of the false positive points (1-specificity). Defined as P( score>=cutpoint | grp=0 ).}
\item{\code{tp}}{Vector of the true positive points (sensitivity). Defined as P( score>=cutpoint | grp=1 ).}
\item{\code{cutpoints}}{Vector of cutpoints corresponding to \code{(fp, tp)} pairs.}
\item{\code{fp.ci}}{Exact CI limits for fp estimates (Clopper and Pearson).}
\item{\code{tp.ci}}{Exact CI limits for tp estimates (Clopper and Pearson).}
\item{\code{data.0}}{Outcome vector for group 0.}
\item{\code{data.1}}{Outcome vector for group 1.}
\item{\code{data.all}}{Dataframe with \code{score} for outcome and \code{status} for group.}
}
}
\description{
This function will create a ROC object that can be used for plotting empirical ROC point or estimating a smooth ROC curve or computing the area under the ROC curve (AUC).
}
\details{
The group with the largest sample mean will be re-labeled as group '1'; group '0' has the smallest sample mean. This orientation often ensures ROC curve will lie above y=x line and the area under the ROC curve will be greater than 0.5. To reverse these default group assignments, set \code{flip.grps=FALSE}.

To accomodate data structured in 2-group format, set \code{as.groups=TRUE}, and define \code{score} to be the outcome from one group and \code{grp} to be the outcome the other. This option is intended for quick and easy usage when there are no covariates data present.

Centering of continuous variables is on by default. This facilitates the interpretation of the default AUC computations, which depend on the (a, b) intercepts.

Set \code{get.ci=FALSE} to improve speed when confidence intervals are not needed.
}
\examples{

## Fake ROC data for examples
y.0 <- round( rnorm(100, mean=0, sd=1)   , 1)
y.1 <- round( rnorm(100, mean=1, sd=1.1) , 1)
age <- rnorm(200, mean=35, sd=8)
size <- sample(c("small", "medium", "big"), 200, replace=TRUE)

## Create an ROC object using 2-group structure
myroc <- set.roc(y.0, y.1, as.groups=TRUE)

## Table of operating points
with(myroc, cbind(cutpoint, fp, fp.ci, tp, tp.ci))

## Fraction of group 1 scores at or above the overall median score
cc <- median(myroc$data.all$score)
with(myroc, cbind(cutpoint, tp, tp.ci))[myroc$cutpoint>cc,][1,]

## Create an ROC object using standard data structure
y <- c(y.0, y.1)
d <- c(rep(0,100),rep(1,100))
x <- data.frame(age, size)

myroc <- set.roc(y, d, x)

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
