% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothGeneric.R
\name{smooth}
\alias{smooth}
\title{Generic function for smoothing ROC curves}
\usage{
smooth(roc.obj, ...)
}
\arguments{
\item{x}{An ROC object.}
}
\value{

}
\description{

}
\details{

}
\examples{

}
\references{
Add references here
}
\seealso{
\url{smooth.default} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{ROC}
\keyword{curve,}
\keyword{density}
\keyword{smooth,}
