% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothRoc.R
\name{smooth.roc}
\alias{smooth.roc}
\title{Smooth an empirical ROC with kernal regression smoothers}
\usage{
\method{smooth}{roc}(roc.obj, adj = 0.8, acc = 0.1, ...)
}
\arguments{
\item{roc.obj}{ROC object from \code{set.roc} function.}

\item{adj}{Bandwith for kernal smoothers (used for both groups). Default is \code{0.8}.}

\item{acc}{Resolution along the x-axis (1-Specificity). Default is \code{0.01}.}
}
\value{
Returns a smoothed ROC curve (list) with the following elements:

\describe{
\item{\code{x}}{Survivor function for group 0.}
\item{\code{y}}{Survivor function for group 1.}
\item{\code{density.0}}{Kernal density estimate for group 0.}
\item{\code{density.1}}{Kernal density estimate for group 1.}
}
}
\description{
This function smooth the empirical ROC curve by smoothing the underlying density estimates for the two group. Kernal regression smoothers are used for the smoothing.
}
\details{
The bandwith for kernal smoothers is equal in the two groups.
}
\examples{

## Fake ROC data for examples 
y.0 <- round( rnorm(100, mean=0, sd=1)   , 1) 
y.1 <- round( rnorm(100, mean=1, sd=1.1) , 1)
myroc <- set.roc(score=y.0, grp=y.1, as.groups=TRUE)

## Let's smooth
myroc.smooth <- smooth(myroc)

## Plotting
plot(myroc)
lines(myroc.smooth, col="purple", lwd=2)
lines(smooth(myroc,adj=0.5), col="hotpink", lwd=2)

## Small bandwidth recovers empirical ROC curve
plot(myroc, roc.line="linear", line.col='red')
lines(smooth(myroc, adj=0.01), col="black")  

}
\references{
Add references here
}
\seealso{
\url{www.statisticalevidence.com} \code{\link{smooth.roc}}
}
\author{
Jeffrey D Blume, \email{j.blume@vanderbilt.edu}
}
\keyword{AUC}
\keyword{ROC}
\keyword{curve,}
