% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{auc}
\alias{auc}
\title{Calculate the area under an ROC curve.}
\usage{
auc(fpr, tpr)
}
\arguments{
\item{fpr}{A numeric vector, representing the false positive rate (FPR)
of points on a ROC curve.}

\item{tpr}{A numeric vector, representing the true positive rate (TPR) of
points on a ROC curve, assumed to be corresponding to those represented
in \code{fpr}.}
}
\value{
The AUC.
}
\description{
Given a receiver operating characteristic (ROC) curve,
calculates the area under said curve (AUC).
}
\details{
Arguments \code{fpr} and \code{tpr} must contain only values between
0 and 1, without any NA or NaN values, and must be the same length.
Otherwise, \code{auc} will throw an appropriate error.
}
\examples{
roc.df = roc(c(1, 2, 3, 4, 4, 5), c(FALSE, FALSE, FALSE, TRUE, FALSE, TRUE))
auc(roc.df$FPR, roc.df$TPR)
}

